Adds folder watcher service
Implements a service that monitors a folder for changes and ingests new or modified files.

The watcher uses a database table to track the state of files,
avoiding redundant processing. It also supports ignoring files
based on glob patterns and moving processed files to a separate directory.

Enables scheduled execution and configuration via properties.